add Modify Date
writer : jieun Lee
upload : 2015.06.12

+++ MoneyBook/src/GUI/deleteDialog.java
@@ -95,7 +95,8 @@ class deleteDialog extends JDialog
 							break;
 						}
 					}
-					hm.put(klist[0],deleteItem);
+					if(deleteItem.size()> 0) hm.put(klist[0],deleteItem);
+					else hm.remove(klist[0]);
 				}
 				setVisible(false);
 				dispose();
